+++ packages/css-tokenizer/.eslintrc.js
@@ -41,6 +41,7 @@ module.exports = {
                 SwitchCase: 1,
             },
         ],
+        '@typescript-eslint/member-delimiter-style': 'error',
         'jsdoc/multiline-blocks': ['error', { noSingleLineBlocks: true }],
         'import/prefer-default-export': 'off',
         'import-newlines/enforce': ['error', { items: 3, 'max-len': MAX_LINE_LENGTH }],