+++ packages/dnr-rulesets/tasks/build.ts
@@ -31,6 +31,7 @@ const createTxt = async (): Promise<void> => {
  * Used as a cleanup step after filters conversion to remove unnecessary txt files.
  *
  * @param dir Directory with txt files.
+ *
  * @returns Promise that resolves when all txt files are removed.
  */
 const removeTxtFiles = async (dir: string): Promise<void> => {
@@ -52,9 +53,10 @@ const removeTxtFiles = async (dir: string): Promise<void> => {
  *     "enabled": true,
  *     "path": "filters/declarative/rules.json"
  *   }]
- * }
+ * }.
  *
- * we should find corresponding text file in resources, and then convert and save json to path specified in the manifest
+ * We should find corresponding text file in resources,
+ * and then convert and save json to path specified in the manifest.
  */
 const build = async (): Promise<void> => {
     const metadata = await startDownload();