+++ packages/examples/tswebextension-mv3/scripts/constants.ts
@@ -11,6 +11,7 @@ export const DEFAULT_EXTENSION_CONFIG: Configuration = {
     staticFiltersIds: [1, 2, 3, 4, 9, 14],
     logLevel: LogLevel.Debug,
     customFilters: [],
+    trustedDomains: [],
     allowlist: [],
     userrules: Object.assign(
         FilterListPreprocessor.createEmptyPreprocessedFilterList(),