+++ packages/tswebextension/src/lib/mv2/background/cosmetic-api.ts
@@ -18,7 +18,7 @@ type ScriptsAndScriptletsDataMv2 = {
     /**
      * JS and scriptlet rules **combined** script text.
      */
-    scriptText: string,
+    scriptText: string;
 };
 
 /**
@@ -94,6 +94,9 @@ export class CosmeticApi extends CosmeticApiCommon {
      * @param frameUrl Frame url.
      *
      * @returns Script text or empty string if no script rules are passed.
+     *
+     * @todo Move to common class when a way to use appContext in common
+     * class will be found.
      */
     public static getScriptText(rules: CosmeticRule[], frameUrl?: string): string {
         const permittedRules = CosmeticApi.sanitizeScriptRules(rules);
@@ -104,15 +107,8 @@ export class CosmeticApi extends CosmeticApiCommon {
 
         const uniqueScriptStrings = new Set<string>();
 
-        let debug = false;
-        const { configuration } = appContext;
-        if (configuration) {
-            const { settings } = configuration;
-            if (settings) {
-                // https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2584
-                debug = settings.debugScriptlets;
-            }
-        }
+        // https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2584
+        const debug = appContext?.configuration?.settings?.debugScriptlets;
 
         const scriptParams = {
             debug,