+++ packages/tswebextension/src/lib/mv3/background/engine-api.ts
@@ -24,12 +24,11 @@ import { isHttpOrWsRequest, isHttpRequest, getHost } from '../../common/utils/ur
 
 import { allowlistApi } from './allowlist-api';
 import { type ConfigurationMV3 } from './configuration';
-import { DocumentApi } from './document-api';
 
 const ASYNC_LOAD_CHINK_SIZE = 5000;
 
 type EngineConfig = Pick<ConfigurationMV3, 'userrules' | 'quickFixesRules' | 'verbose'> & {
-    filters: IFilter[],
+    filters: IFilter[];
 };
 
 /**
@@ -193,13 +192,6 @@ export class EngineApi {
         }
 
         const frameUrl = getHost(url);
-
-        // Checks if an allowlist rule exists at the document level,
-        // then discards all cosmetic rules.
-        const allowlistFrameRule = DocumentApi.matchFrame(url);
-        if (allowlistFrameRule) {
-            return new CosmeticResult();
-        }
 
         const request = new Request(url, frameUrl, RequestType.Document);