+++ packages/tswebextension/src/lib/mv3/tabs/frame.ts
@@ -6,11 +6,21 @@ import { type PreparedCosmeticResultCommon, FrameCommon } from '../../common/tab
  * Prepared cosmetic result.
  * This type represents the processed cosmetic data extracted from the initial cosmetic result.
  */
-type PreparedCosmeticResultMV3 = PreparedCosmeticResultCommon & {
+type GeneralPreparedCosmeticResultMV3 = PreparedCosmeticResultCommon & {
+    /**
+     * CSS styles extracted from the cosmetic result.
+     */
+    cssText?: string;
+};
+
+/**
+ * Prepared cosmetic result for MV3 with scriptlets data and script texts.
+ */
+type PreparedCosmeticResultMV3 = GeneralPreparedCosmeticResultMV3 & {
     /**
      * Script texts extracted from the cosmetic result.
      */
-    scriptTexts: string[],
+    scriptTexts: string[];
 
     /**
      * A list of scriptlet data extracted from the cosmetic result.
@@ -18,9 +28,30 @@ type PreparedCosmeticResultMV3 = PreparedCosmeticResultCommon & {
     scriptletDataList: ScriptletData[];
 
     /**
-     * CSS styles extracted from the cosmetic result.
+     * Using never here ensures this type cannot have a scriptText property.
      */
-    cssText?: string;
+    scriptText?: never;
+};
+
+/**
+ * Prepared cosmetic result for MV3 for user scripts API with already combined
+ * script text.
+ */
+type PreparedCosmeticResultMV3ForUserScripts = GeneralPreparedCosmeticResultMV3 & {
+    /**
+     * Script text extracted from the cosmetic result.
+     */
+    scriptText: string;
+
+    /**
+     * Using never here ensures this type cannot have scriptTexts and scriptletDataList properties.
+     */
+    scriptTexts?: never;
+
+    /**
+     * Using never here ensures this type cannot have scriptTexts and scriptletDataList properties.
+     */
+    scriptletDataList?: never;
 };
 
 /**
@@ -34,5 +65,5 @@ export class FrameMV3 extends FrameCommon {
      *
      * This data is saved in the frame because it is needed for injecting cosmetic rules into the frames.
      */
-    public preparedCosmeticResult?: PreparedCosmeticResultMV3;
+    public preparedCosmeticResult?: PreparedCosmeticResultMV3 | PreparedCosmeticResultMV3ForUserScripts;
 }