+++ packages/tswebextension/test/lib/mv3/background/dynamic-rules-api.test.ts
@@ -7,7 +7,12 @@ import {
 import browser from 'webextension-polyfill';
 
 import DynamicRulesApi from '../../../../src/lib/mv3/background/dynamic-rules-api';
-import { ALLOWLIST_FILTER_ID, QUICK_FIXES_FILTER_ID, USER_FILTER_ID } from '../../../../src/lib/common/constants';
+import {
+    ALLOWLIST_FILTER_ID,
+    BLOCKING_TRUSTED_FILTER_ID,
+    QUICK_FIXES_FILTER_ID,
+    USER_FILTER_ID,
+} from '../../../../src/lib/common/constants';
 import { createFilter } from '../helpers';
 
 describe('DynamicRulesApi', () => {
@@ -34,6 +39,7 @@ describe('DynamicRulesApi', () => {
             let conversionResult = await DynamicRulesApi.updateDynamicFiltering(
                 createFilter([quickFixesRule], QUICK_FIXES_FILTER_ID),
                 createFilter([allowlistRule], ALLOWLIST_FILTER_ID),
+                createFilter([], BLOCKING_TRUSTED_FILTER_ID),
                 createFilter([userRule], USER_FILTER_ID),
                 [createFilter([customRule], 1000)],
                 [],
@@ -50,6 +56,7 @@ describe('DynamicRulesApi', () => {
             conversionResult = await DynamicRulesApi.updateDynamicFiltering(
                 createFilter([quickFixesRule], QUICK_FIXES_FILTER_ID),
                 createFilter([allowlistRule], ALLOWLIST_FILTER_ID),
+                createFilter([], BLOCKING_TRUSTED_FILTER_ID),
                 createFilter([userRule], USER_FILTER_ID),
                 [createFilter([customRule], 1000)],
                 [],
@@ -69,6 +76,7 @@ describe('DynamicRulesApi', () => {
             conversionResult = await DynamicRulesApi.updateDynamicFiltering(
                 createFilter([quickFixesRule], QUICK_FIXES_FILTER_ID),
                 createFilter([allowlistRule], ALLOWLIST_FILTER_ID),
+                createFilter([], BLOCKING_TRUSTED_FILTER_ID),
                 createFilter([userRule], USER_FILTER_ID),
                 [createFilter([customRule], 1000)],
                 [],
@@ -89,6 +97,7 @@ describe('DynamicRulesApi', () => {
             conversionResult = await DynamicRulesApi.updateDynamicFiltering(
                 createFilter([quickFixesRule], QUICK_FIXES_FILTER_ID),
                 createFilter([allowlistRule], ALLOWLIST_FILTER_ID),
+                createFilter([], BLOCKING_TRUSTED_FILTER_ID),
                 createFilter([userRule], USER_FILTER_ID),
                 [createFilter([customRule], 1000)],
                 [],