+++ packages/tswebextension/CHANGELOG.md
@@ -5,13 +5,41 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.1.1] - 2025-05-29
+## Unreleased <!-- release/browser-extension-v5.2 -->
+
+### Added
+
+- `filterId` query param to `documentBlockingPageUrl` for MV2.
+- Support for Chrome's User Scripts API that allows more reliable script
+  injection in MV3 if developer mode is enabled.
+- Static getter `isUserScriptsApiSupported` to check if the current browser
+  supports User Scripts API.
+
+## Changed
+
+<!-- FIXME: update dependencies to the latest versions -->
+- Updated [@adguard/agtree] to `v3.1.5`.
+- Updated [@adguard/scriptlets] to `v2.2.1`.
+- Updated [@adguard/tsurlfilter] to `v3.x.x`.
 
 ### Fixed
 
+- Invalid HTML rule selectors are breaking site loading
+  [AdguardBrowserExtension#2646], [AdguardBrowserExtension#2826].
+- Scriptlet rules are not displayed in the filtering log [AdguardBrowserExtension#3164].
+- Stealth mode's `Hide Referer from third parties` option may break some websites [AdguardBrowserExtension#2839].
+- Filtering doesn't work in Edge's split screen [AdguardBrowserExtension#2832].
+- `$replace` rules may break some websites [AdguardBrowserExtension#3122].
+- Increase file size limit to 10MB for `$replace` rules in Firefox [AdguardBrowserExtension#3192]
 - CLI was not worked since to ESM builds.
 
-[3.1.1]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v3.1.1
+[AdguardBrowserExtension#2646]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2646
+[AdguardBrowserExtension#2826]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2826
+[AdguardBrowserExtension#2832]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2832
+[AdguardBrowserExtension#2839]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2839
+[AdguardBrowserExtension#3122]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3122
+[AdguardBrowserExtension#3164]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3164
+[AdguardBrowserExtension#3192]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3192
 
 ## [3.1.0] - 2025-05-28
 
@@ -1189,7 +1217,7 @@ its setting [AdguardBrowserExtension#2584].
 
 ## [0.1.1] - 2023-04-04
 
-## Changed
+### Changed
 
 - Improved injection algorithm for cosmetic rules (js and css).
   logic using the Finite State Machine to avoid double injections with